% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\name{delete_topic}
\alias{delete_topic}
\title{Delete a topic}
\usage{
delete_topic(topic, ...)
}
\arguments{
\item{topic}{A character string containing an SNS Topic Amazon Resource Name
(ARN).}

\item{...}{Additional arguments passed to \code{\link{snsHTTP}}.}
}
\value{
If successful, a logical \code{TRUE}. Otherwise, a data structure of
class \dQuote{aws_error} containing any error message(s) from AWS and
information about the request attempt.
}
\description{
Delete an SNS topic
}
\details{
Delete an existing topic.
}
\author{
Thomas J. Leeper
}
\references{
\href{http://docs.aws.amazon.com/sns/latest/api/API_CreateTopic.html}{CreateTopic}
}
\seealso{
\code{link{create_topic}}
}

