\name{aws.sns-package}
\alias{aws.sns-package}
\alias{aws.sns}
\docType{package}
\title{AWS SNS Client Package}
\description{A simple client package for the Amazon Web Services (AWS) Simple Notification Service (SNS) REST API.}
%\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
