% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.sns-package}
\alias{aws.sns-package}
\alias{aws.sns}
\title{aws.sqs}
\description{
AWS SNS Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) Simple Notification Service (SNS) REST API.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
